+++ PROGRESS.md
@@ -7,6 +7,7 @@ Entry format: `- (YYYY-MM-DD) Description of feature-level change (PRD #X, miles
 ## [Unreleased]
 
 ### Added
+- (2026-04-27) Extracted the yt-dlp video download function from the micro.blog module into a standalone shared utility (`src/video-download.js`). Previously `downloadShortVideo()` was a private function inside `post-microblog.js`; it now lives in its own module and is imported back by micro.blog. This is the foundation for the upcoming video embed feature — Bluesky, Mastodon, and LinkedIn platform posters will all import from the same shared download module rather than duplicating the yt-dlp logic. All 20 existing micro.blog tests pass.
 - (2026-04-27) Extended the content scanner to cover three additional shows: Datadog Illuminated, Enlightning, and You Choose. Previously only Thunder and Software Defined Interviews were auto-detected; episodes from the other shows had to be added to the spreadsheet by hand. The scanner now fetches all five YouTube playlists in parallel and deduplicates against existing spreadsheet URLs. The daily GitHub Actions workflow now runs the scan automatically before the post-priority step, so the spreadsheet stays current without manual intervention. 19 unit tests cover playlist ID usage, show name labels, private video filtering, URL format, and pagination. 213 tests passing total.
 - (2026-04-25) Launched the dynamic About page: `updateAboutPage()` reads the live production spreadsheet, determines which content channels are currently active based on freshness thresholds, and pushes updated content to the Micro.blog About page via XML-RPC — only when the content has actually changed. The page now shows a photo+bio header (HTML flex layout) followed by a flat list of active channel links. Channels drop off automatically when they go quiet; SDI is always last. Discovered and fixed three Micro.blog API quirks during live testing: `getPages` requires the username as blogId (not integer 1), the page identifier field is `id` not `pageID`, and integers in responses use `<i4>` not `<int>`. 190 tests passing.
 - (2026-04-25) Added channel config and activity detection for the dynamic About page: `src/config/about-page-channels.js` defines which content channels (Thunder, Enlightning, Datadog Illuminated, You Choose, Conference Talks, GitHub, SDI) appear based on freshness thresholds, and `getActiveChannels()` in `src/update-about-page.js` filters spreadsheet rows by type and show name to determine which channels have published recently enough to appear. 22 unit tests cover all activity scenarios. Channels sort by most recent content date; SDI always appears last regardless of activity.
