+++ journal/entries/2026-04/2026-04-25.md
@@ -1210,3 +1210,29 @@ No significant technical decisions documented for this development session
 
 ═══════════════════════════════════════
 
+
+
+## 6:09:27 PM CDT - Commit: f343839d - journal: 2026-04-25 session notes [skip ci]
+
+### Summary - f343839d
+
+The developer documented session notes for April 25, 2026, in the journal entry file. The update included details about a merge from the main branch and the resolution of conflicts in the journal and PRD files.
+
+### Development Dialogue - f343839d
+
+No significant dialogue found for this development session
+
+### Technical Decisions - f343839d
+
+No significant technical decisions documented for this development session
+
+### Commit Details - f343839d
+
+**Files Changed**:
+- journal/entries/2026-04/2026-04-25.md
+
+**Lines Changed**: ~22 lines
+**Message**: "journal: 2026-04-25 session notes [skip ci]"
+
+═══════════════════════════════════════
+
