+++ journal/entries/2026-04/2026-04-27.md
@@ -110,3 +110,30 @@ No significant technical decisions documented for this development session
 
 ═══════════════════════════════════════
 
+
+
+## 6:41:29 AM CDT - Commit: 285c44f6 - chore: archive PRD #36 to done [skip ci]
+
+### Summary - 285c44f6
+
+The commit archived PRD #36 by deleting the corresponding documentation file and creating a new archive in the done directory. This action is part of the routine maintenance to manage project documentation effectively.
+
+### Development Dialogue - 285c44f6
+
+No significant dialogue found for this development session
+
+### Technical Decisions - 285c44f6
+
+No significant technical decisions documented for this development session
+
+### Commit Details - 285c44f6
+
+**Files Changed**:
+- prds/36-extend-scan-new-content.md
+- prds/done/36-extend-scan-new-content.md
+
+**Lines Changed**: ~244 lines
+**Message**: "chore: archive PRD #36 to done [skip ci]"
+
+═══════════════════════════════════════
+
